Moved type-patches.ts to injections/type-patches.ts
