+++ src/index.ts
@@ -1,3 +1,6 @@
+export * from './types/index.js'
+export * from './injections/index.js'
+export * from './commands/index.js'
 export * from './config.js'
 export * from './conflict-resolver.js'
 export * from './constants.js'
@@ -11,8 +14,5 @@ export * from './module-loader.js'
 export * from './symtable.js'
 export * from './template-processor.js'
 export * from './transformation.js'
-export * from './type-patches.js'
 export * from './utils.js'
 export { default as templateProcessor } from './template-processor.js'
-export * from './types/index.js'
-export * from './commands/index.js'
