+++ src/injections/index.ts
@@ -1,1 +1,2 @@
 export * from './classes.js'
+export * from './type-patches.js'
